UPDATE, FIX: Backend - Solicitudes complete updated, Solicitud migration fixed

+++ backend/laravel_src/database/migrations/2021_03_03_184655_create_parte_solicitud_table.php
@@ -24,7 +24,7 @@ class CreateParteSolicitudTable extends Migration
             $table->float('margen')->nullable();
             $table->float('peso')->nullable();
             $table->float('flete')->nullable();
-            $table->float('monto')->nullable();
+            $table->double('monto', 10, 2)->nullable();
             $table->boolean('backorder')->default(false);
             $table->timestamps();
